refactor(zjcrypto): 统一使用 #nosec 注释替代 nolint:gosec
style: 修复 archive.go 中的缩进问题
feat(format): 将 FilenameLen 类型从 byte 改为 uint16 避免溢出
ci(release): 为每个发布文件生成单独的校验文件
chore: 移除不再需要的 .gosec 配置文件

+++ internal/format/header_test.go
@@ -353,7 +353,7 @@ func TestMarshalBinaryOptimizedConsistency(t *testing.T) {
 				Magic:       [4]byte{'F', 'Z', 'J', 0x01},
 				Version:     0x0100,
 				Algorithm:   0x02,
-				FilenameLen: byte(len(tc.filename)),
+				FilenameLen: uint16(len(tc.filename)),
 				Filename:    tc.filename,
 				FileSize:    1024,
 				// #nosec G115 - 测试数据，不会溢出
@@ -536,7 +536,7 @@ func TestMarshalBinaryWithRealData(t *testing.T) {
 		Version:     0x0100,
 		Algorithm:   0x02,
 		Flags:       0x00,
-		FilenameLen: byte(len(filename)),
+		FilenameLen: uint16(len(filename)),
 		FileSize:    10485760,
 		// #nosec G115 - 测试数据，不会溢出
 		Timestamp:   uint32(time.Now().Unix()),
